backend/requirements.txt
eventlet==0.24.1
gunicorn[eventlet]==19.9.0
Flask==1.0.2
Flask-SQLAlchemy==2.3.2
Flask-Cors==3.0.7
Flask-Login==0.4.1
Flask-Restplus==0.12.1
Flask-Restplus-Patched==0.1.10
Flask-Migrate== 2.4.0
Flask-Socketio==3.3.2
Flask-Marshmallow==0.7.0
SQLAlchemy-Utils==0.33.11
Marshmallow-sqlalchemy==0.16.2
bcrypt==3.1.3
passlib==1.7.1
